reorganizing docs folder
